Type the parser boundary, and flatten the type mapping table
The loader had 29 T.untyped, one per openapi3_parser node, because the
gem ships no type information so the file where the model is built was
the one file Sorbet could not check. A shim now declares the node surface
the loader actually touches, typed from the gem's own documented return
values, and the loader names real node classes. Three T.untyped remain,
and all three are genuinely untyped document data.

Sorbet immediately found two bugs:

- multipleOf is any positive number in OpenAPI, but IntegerSchema stored
  it as an Integer, so an integer schema with multipleOf: 0.5 crashed the
  generator with a T::Struct TypeError.
- A discriminated union member that is an inline schema has no name to
  map a discriminator value onto, so it was silently dropped from the
  mapping and left undispatchable. It is refused now.

Node collections stay out of the generator's own sigs: the RBI declares
them generic for inference, but sorbet-runtime evaluates real sigs, and
the classes are not generic at runtime. Each method reads the collection
off its parent node instead, which is the better shape regardless.

Cycle detection moved out of alias expansion. A cycle is a property of
the type graph, so it is walked once at construction, deterministically
from the lexicographically first alias; while_expanding and its three
call sites are gone.

DEFAULT_TYPE_MAPPINGS is one flat literal table rather than a function
with locals and a merge.

+++ lib/oapi/codegen/model/schema.rb
@@ -57,7 +57,7 @@ module Oapi
         const :maximum, T.nilable(Integer), default: nil
         const :exclusive_minimum, T::Boolean, default: false
         const :exclusive_maximum, T::Boolean, default: false
-        const :multiple_of, T.nilable(Integer), default: nil
+        const :multiple_of, T.nilable(Numeric), default: nil
         const :meta, Meta, factory: -> { Meta.new }
       end
